Make history items collapsible and enforce coverage threshold
