+++ history.js
@@ -1,4 +1,5 @@
 import { loadHistory } from './historyManager.js';
+import { renderMarkdown } from './render.js';
 
 document.addEventListener('DOMContentLoaded', () => {
   const list = document.getElementById('historyList');
@@ -11,14 +12,19 @@ document.addEventListener('DOMContentLoaded', () => {
   }
   for (const item of items) {
     const li = document.createElement('li');
+    const details = document.createElement('details');
+    const summary = document.createElement('summary');
     const titleLink = document.createElement('a');
     titleLink.href = item.url;
     titleLink.textContent = item.title;
     titleLink.target = '_blank';
-    li.appendChild(titleLink);
-    const info = document.createElement('div');
-    info.textContent = `${item.channel} – ${item.summary}`;
-    li.appendChild(info);
+    summary.appendChild(titleLink);
+    summary.append(` – ${item.channel}`);
+    details.appendChild(summary);
+    const content = document.createElement('div');
+    content.innerHTML = renderMarkdown(item.summary);
+    details.appendChild(content);
+    li.appendChild(details);
     list.appendChild(li);
   }
 });
