+++ package.json
@@ -7,7 +7,7 @@
     "test": "tests"
   },
   "scripts": {
-    "test": "node --test"
+    "test": "c8 --check-coverage --lines 85 --functions 85 --statements 85 node --test"
   },
   "keywords": [],
   "author": "",
